[fix] utile 폴더 수정 - server 폴더 생성 : server에서 사용되는 함수 묶음 정리(getImage) - client 폴더 생성 : client에서 사용되는 함수 묶음 정리(tabbableSelector) - shard 폴더 생성 : 공통 사용 함수 묶음 정리(tw, wait, measureTime)
