+++ src/utils/client/index.ts
@@ -1,0 +1,2 @@
+// src/utils/client/index.ts
+export { tabbableSelector } from './tabbable'
