+++ src/utils/client/tabbable.ts
@@ -1,0 +1,16 @@
+// src/utils/client/tabbable.ts
+export const tabbableSelector = [
+  'a[href]',
+  'area[href]',
+  'button',
+  'input',
+  'select',
+  'textarea',
+  'iframe',
+  'summary',
+  'details',
+  'video[controls]',
+  'audio[controls]',
+  '[contenteditable]:not([contenteditable="false"])',
+  '[tabindex]:not([tabindex="-1"])',
+].join(',')
