--- src/utils/index.ts
@@ -1,4 +1,0 @@
-export { default as measureTime } from './measure-time'
-export { default as getImage } from './get-image'
-export { default as wait } from './wait'
-export { default as tw } from './tw'
