+++ src/utils/server/index.ts
@@ -1,0 +1,2 @@
+// src/utils/server/index.ts
+export { getImage } from './get-image'
