+++ src/utils/shared/measure-time.ts
@@ -1,7 +1,5 @@
-export default async function measureTime(
-  label: string,
-  fn: () => Promise<void>
-) {
+// src/utils/shared/measure-time.ts
+export async function measureTime(label: string, fn: () => Promise<void>) {
   const start = Date.now()
   await fn()
   const end = Date.now()
