+++ src/utils/shared/tw.ts
@@ -1,6 +1,6 @@
 import { type ClassValue, clsx } from 'clsx'
 import { twMerge } from 'tailwind-merge'
 
-export default function tw(...classes: ClassValue[]) {
+export function tw(...classes: ClassValue[]) {
   return twMerge(clsx(...classes))
 }
